Revert "Debugging texturing"
This reverts commit 2d827688afa16064a3d3fc6f456ae42342bb37eb.

+++ terra-meshgen/src/main/java/com/ritualsoftheold/terra/mesher/VoxelMesher.java
@@ -31,6 +31,6 @@ public interface VoxelMesher {
     
     IntList getIndices();
     
-    IntList getTextureCoords();
+    FloatList getTextureCoords();
 
 }
